+++ Modumate/Source/Modumate/Public/UI/PendingAngleActor.h
@@ -20,5 +20,5 @@ public:
 public:
 	FVector WorldPosition;
 	FVector WorldDirection;
-	float CurrentAngle;
+	double CurrentAngle;
 };
